requirements.txt
matplotlib==3.7.5
numpy==1.24.4
torch==2.4.1
torchvision==0.19.1
tqdm==4.66.5
pyyaml==6.0.2
